+++ samples/dotnet/kernel-syntax-examples/Example08_RetryHandler.cs
@@ -45,7 +45,7 @@ public static class Example08_RetryHandler
         // Add 401 to the list of retryable status codes
         // Typically 401 would not be something we retry but for demonstration
         // purposes we are doing so as it's easy to trigger when using an invalid key.
-        kernelBuilder = kernelBuilder.Configure(c => c.DefaultHttpRetryConfig.RetryableStatusCodes.Add(System.Net.HttpStatusCode.Unauthorized));
+        kernelBuilder = kernelBuilder.Configure(c => c.DefaultHttpRetryConfig.RetryableStatusCodes.Add((int)System.Net.HttpStatusCode.Unauthorized));
 
         // OpenAI settings - you can set the OPENAI_API_KEY to an invalid value to see the retry policy in play
         kernelBuilder = kernelBuilder.Configure(c => c.AddOpenAICompletion("text-davinci-003", "text-davinci-003", "BAD_KEY"));
